Preprocessor directives for addons
This patch augments the XML dumps with a 'directivelist'
subtree which lists all raw preprocessor directives met
while reading the source code in each configuration.

Also, the addons/cppcheckdata.py file has been extended
to give easy access to the list of directives.

Finally, an new addon, addons/y2038.py, is created to
detect occurrences of glibc symbols when they may not be
Y2038-proof. This is determined for each symbol by checking
whether _TIME_BITS is currently defined equal to 64 when
the symbol is met.